refactor:code splitted into reducers,constants,services and utilities for simplified understanding
