+++ src/constants/AuthConstants.js
@@ -1,0 +1,15 @@
+const initialsignInData = {
+    email: "",
+    password: ""
+}
+const initialSignupData = {
+    firstName: "",
+    lastName: "",
+    email: "",
+    password: ""
+}
+const guestCredentials = {
+    email: "himmatsingh@gmail.com",
+    password: "specialops1.5"
+}
+export { initialsignInData, initialSignupData, guestCredentials };
