+++ src/constants/index.js
@@ -1,0 +1,1 @@
+export { initialsignInData, initialSignupData, guestCredentials } from "./AuthConstants";
