+++ src/context/AuthContext.jsx
@@ -1,27 +1,9 @@
-/* eslint-disable no-unused-vars */
-import React, { createContext, useContext, useState, useEffect, useReducer } from "react";
-import axios from "axios";
+import React, { createContext, useContext, useReducer } from "react";
+import { authReducer, initialAuthState } from "reducers";
 const AuthenticationContext = createContext(null);
 const useAuthentication = () => useContext(AuthenticationContext);
 
 const AuthProvider = ({ children }) => {
-    const initialAuthState = {
-        signinStatus: false,
-        authenticationToken: ""
-    }
-
-    const authReducer = (authStateProvided, authAction) => {
-        const { type, payload } = authAction;
-        switch (type) {
-            case "SIGN-IN":
-                return { ...authStateProvided, ...payload.foundUser, authenticationToken: payload.encodedToken, signinStatus: true }
-            case "SIGN-OUT":
-                return { ...initialAuthState }
-            case "SIGN-UP":
-                return { ...authStateProvided, ...payload.createdUser, authenticationToken: payload.encodedToken, signinStatus: true }
-        }
-    }
-
     const [authState, authDispatch] = useReducer(authReducer, initialAuthState);
     return (
         <AuthenticationContext.Provider value={{ authState, authDispatch }}>
