+++ src/pages/Authentication/SignIn.jsx
@@ -1,43 +1,20 @@
 import React, { useState } from 'react';
 import { Link, useNavigate } from "react-router-dom";
 import { FaChevronRight } from "react-icons/fa";
+import { initialsignInData, guestCredentials } from 'constants';
 import { useAuthentication } from 'context';
+import { signInActionHandler } from 'services';
 import "./authentication.css";
-import axios from 'axios';
+
 const SignIn = () => {
     const navigate = useNavigate();
     const { authDispatch } = useAuthentication();
-    const initialsignInData = {
-        email: "",
-        password: ""
-    }
-    const guestCredentials = {
-        email: "himmatsingh@gmail.com",
-        password: "specialops1.5"
-    }
     const [signinData, setSigninData] = useState(initialsignInData);
     const signInDataHandler = (e) => {
         const { name, value } = e.target;
         setSigninData(prevSigninData => ({ ...prevSigninData, [name]: value }))
     }
 
-    const signInActionHandler = (e, signinDataProvided) => {
-        e.preventDefault();
-        (async () => {
-            try {
-                const { data, status } = await axios.post("/api/auth/login", {
-                    ...signinDataProvided
-                })
-                if (status === 200) {
-                    authDispatch({ type: "SIGN-IN", payload: data });
-                    localStorage.setItem("authenticationToken", data.encodedToken);
-                    navigate("/products")
-                }
-            } catch {
-                console.error("LOGIN NOT POSSIBLE")
-            }
-        })()
-    }
     return (
         <section className="main-content flex-column flex-center">
             <div
@@ -46,7 +23,7 @@ const SignIn = () => {
                 <div className="form-name text-center">
                     <h3>SIGN IN</h3>
                 </div>
-                <form className="gentle-form-group gentle-flex-gap txt-sm" onSubmit={(e) => { signInActionHandler(e, signinData) }}>
+                <form className="gentle-form-group gentle-flex-gap txt-sm" onSubmit={(e) => { signInActionHandler(e, signinData, authDispatch, navigate) }}>
                     <div className="gentle-input-group">
                         <label className="gentle-input-label" htmlFor="user-email">Email Address</label>
                         <input
@@ -101,8 +78,7 @@ const SignIn = () => {
                         </button>
                         <button
                             onClick={(e) => {
-                                e.preventDefault();
-                                signInActionHandler(e, guestCredentials)
+                                signInActionHandler(e, guestCredentials, authDispatch, navigate)
                             }}
                             className="btn btn-info btn-info-hover submit-btn"
                         >
