+++ src/reducers/index.js
@@ -1,0 +1,2 @@
+export { initialAuthState, authReducer } from "./auth-reducer";
+export { filterReducer, initialFilterState } from "./filter-reducer";
