+++ src/services/Authentication/signOutService.js
@@ -1,0 +1,6 @@
+const signOutService = (authDispatch, navigatorHelper) => {
+    authDispatch({ type: "SIGN-OUT" })
+    localStorage.removeItem("authenticationToken");
+    navigatorHelper("/");
+}
+export {signOutService};
