+++ src/services/index.js
@@ -1,0 +1,2 @@
+export { loadProductsAndCategories } from "./Products"
+export { signOutService, signInActionHandler, signupHandler } from "./Authentication";
